task(project): update to 1.21.5 and bump project version to 1.0.4

+++ src/main/java/uk/co/willpoulson/willsfarmingtweaks/mixin/FarmlandBlockMixin.java
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(FarmlandBlock.class)
 public class FarmlandBlockMixin {
 	@Inject(method = "onLandedUpon", at = @At("HEAD"), cancellable = true)
-	private void preventFarmlandTrample(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance, CallbackInfo ci) {
+	private void preventFarmlandTrample(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance, CallbackInfo ci) {
 		ci.cancel();
 	}
 }
